test-requirements.txt
pytest
numpy
pandas
# scikit-learn==1.2.0
# lightgbm==3.3.2
matplotlib
shap
seaborn
# jinja2==3.1.2
ipython
# psutil== 5.9.0
# python-dotenv==1.0.0
streamlit
matplotlib --only-binary :all:
plotly
fastapi
uvicorn
databases
pydantic
sqlalchemy
streamlit-cookies-manager
fastapi[all]
python-jose
passlib
bcrypt==4.0.1
prometheus_fastapi_instrumentator
aiocache
